Updating to latest FlutterFlow output.

+++ lib/backend/supabase/database/tables/pagos.dart
@@ -23,11 +23,11 @@ class PagosRow extends SupabaseDataRow {
   int? get productoId => getField<int>('productoId');
   set productoId(int? value) => setField<int>('productoId', value);
 
-  int? get precioInicial => getField<int>('precioInicial');
-  set precioInicial(int? value) => setField<int>('precioInicial', value);
+  double? get precioInicial => getField<double>('precioInicial');
+  set precioInicial(double? value) => setField<double>('precioInicial', value);
 
-  int? get precioFinal => getField<int>('precioFinal');
-  set precioFinal(int? value) => setField<int>('precioFinal', value);
+  double? get precioFinal => getField<double>('precioFinal');
+  set precioFinal(double? value) => setField<double>('precioFinal', value);
 
   String? get status => getField<String>('status');
   set status(String? value) => setField<String>('status', value);
